[xtext generator] Extracted the composition of the generated validator classes' names from 'ValidatorFragment2' to 'ValidatorNaming', updated 'QuickfixProviderFragment2' correspondingly.
