+++ plugins/org.eclipse.xtext.xtext.generator/xtend-gen/org/eclipse/xtext/xtext/generator/validation/ValidatorNaming.java
@@ -1,0 +1,47 @@
+/**
+ * Copyright (c) 2015 itemis AG (http://www.itemis.eu) and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+package org.eclipse.xtext.xtext.generator.validation;
+
+import com.google.inject.Inject;
+import org.eclipse.xtext.Grammar;
+import org.eclipse.xtext.GrammarUtil;
+import org.eclipse.xtext.xbase.lib.Extension;
+import org.eclipse.xtext.xtext.generator.XtextGeneratorNaming;
+import org.eclipse.xtext.xtext.generator.model.TypeReference;
+
+/**
+ * Separates the composition of the generated validator classes' names from
+ *  the template of those classes (which may be specialized by clients),
+ *  since the name is used in the template of the generated quickfix provider classes, too.
+ * 
+ * @author Christian Schneider - Initial contribution and API
+ */
+@SuppressWarnings("all")
+public class ValidatorNaming {
+  @Inject
+  @Extension
+  private XtextGeneratorNaming _xtextGeneratorNaming;
+  
+  public TypeReference getValidatorClass(final Grammar grammar) {
+    String _runtimeBasePackage = this._xtextGeneratorNaming.getRuntimeBasePackage(grammar);
+    String _plus = (_runtimeBasePackage + ".validation.");
+    String _simpleName = GrammarUtil.getSimpleName(grammar);
+    String _plus_1 = (_plus + _simpleName);
+    String _plus_2 = (_plus_1 + "Validator");
+    return new TypeReference(_plus_2);
+  }
+  
+  protected TypeReference getAbstractValidatorClass(final Grammar grammar) {
+    String _runtimeBasePackage = this._xtextGeneratorNaming.getRuntimeBasePackage(grammar);
+    String _plus = (_runtimeBasePackage + ".validation.Abstract");
+    String _simpleName = GrammarUtil.getSimpleName(grammar);
+    String _plus_1 = (_plus + _simpleName);
+    String _plus_2 = (_plus_1 + "Validator");
+    return new TypeReference(_plus_2);
+  }
+}
